+++ R/qsm_summary.R
@@ -35,8 +35,84 @@
 qsm_summary <- function(cylinder, radius = "modified", triangulation = FALSE) {
   message("Creating QSM Summary")
 
+  # rTwig ----------------------------------------------------------------------
+  if (all(c("id", "parent", "start_x", "branch_order") %in% colnames(cylinder))) {
+    # Setup Radius
+    if (radius == "modified") {
+      cylinder$sumrad <- cylinder$radius
+    } else if (radius == "unmodified") {
+      cylinder$sumrad <- cylinder$raw_radius
+    } else if (radius == "old") {
+      cylinder$sumrad <- cylinder$raw_radius
+    }
+
+    dbh <- cylinder %>%
+      filter(.data$branch_order == 0 & .data$branch == 1) %>%
+      arrange(.data$branch_position) %>%
+      select(.data$length, .data$sumrad)
+
+    # Finds the DBH cylinder
+    for (i in 1:nrow(dbh)) {
+      DBHCyl <- sum(dbh$length[1:i])
+      if (DBHCyl >= 1.37) {
+        break
+      }
+    }
+
+    DBHCyl <- as.numeric(i)
+
+    dbh_qsm_cm <- dbh$sumrad[DBHCyl] * 200
+
+    tree_height_m <- max(cylinder$start_z) - min(cylinder$start_z)
+
+    # Branch Order Summary
+    summary <- cylinder %>%
+      mutate(
+        Volume = pi * .data$sumrad^2 * .data$length,
+        SurfaceArea = 2 * pi * .data$sumrad * .data$length
+      ) %>%
+      group_by("branch_order") %>%
+      summarize(
+        tree_volume_L = sum(.data$Volume, na.rm = TRUE) * 1e3,
+        tree_area_m2 = sum(.data$SurfaceArea, na.rm = TRUE)
+      )
+
+    # Total Summary
+    tree_volume_L <- summary %>%
+      summarize(tree_volume_L = sum(.data$tree_volume_L))
+
+    stem_volume_L <- summary %>%
+      filter(.data$branch_order == 0) %>%
+      summarize(stem_volume_L = sum(.data$tree_volume_L))
+
+    branch_volume_L <- summary %>%
+      filter(.data$branch_order != 0) %>%
+      summarize(branch_volume_L = sum(.data$tree_volume_L))
+
+    tree_area_m2 <- summary %>%
+      summarize(tree_area_m2 = sum(.data$tree_area_m2))
+
+    stem_area_m2 <- summary %>%
+      filter(.data$branch_order == 0) %>%
+      summarize(stem_area_m2 = sum(.data$tree_area_m2))
+
+    branch_area_m2 <- summary %>%
+      filter(.data$branch_order != 0) %>%
+      summarize(branch_area_m2 = sum(.data$tree_area_m2))
+
+    summary2 <- bind_cols(
+      "dbh_qsm_cm" = dbh_qsm_cm,
+      "tree_height_m" = tree_height_m,
+      stem_volume_L,
+      branch_volume_L,
+      tree_volume_L,
+      stem_area_m2,
+      branch_area_m2,
+      tree_area_m2
+    )
+  }
   # TreeQSM --------------------------------------------------------------------
-  if (all(c("parent", "extension", "branch", "BranchOrder") %in% colnames(cylinder))) {
+  else if (all(c("parent", "extension", "branch", "BranchOrder") %in% colnames(cylinder))) {
     # Setup Radius
     if (radius == "modified") {
       cylinder$sumrad <- cylinder$radius